fix(pwa): update logic to enable "start game" button
Context:
When all players are selected and there's no court to start a new game
at, it's not possible to start a game and the "Start Game" button is
disabled. After that, when a court becomes available, the button does
not react to that and stays disabled.

Changes:
Ensure the logic takes in account of when a court becomes available.