fix(pwa): fix player index duplicate
Context:
When adding a new player, the index of this player may be the same as
the last player in the list. This affects setting player states when
starting or finishing a game.

In addition, update icon's size from fixed unit of size at `1rem` to relative unit of size at `1em`.